Project 003 - Discord AI ChatBot/requirements.txt
python_dotenv
groq
pillow
Ipython
pypdf2